Create client model
This will create a client instance. This will be our dummy user with little to no security.

They have a name that is specified.

The id and key (api key) are generated.

The campaigns, templates, populations, suppressionLists, archives are all generated by the url of the settings. If the settings is in development then we set it to localhost. If the settings are in production we will set them to the azure url.