Create population view and post request
This will make handle a POST request for our application. If the population parameters are correct the model will be saved.

However, if the population instance not correct due to duplicate names in the db, a 409 response is given which is in accordance with Dispatch.

If another error arises, then a 400 response is given which is also in accordance with Dispatch.